+++ src/common/metadata-utils/index.ts
@@ -385,16 +385,14 @@ class MetadataUtils {
   }
 
   // Get default org that is currently selected for user
-  public static async getCurrentOrg(type = "any") {
-    const displayOrgCommand = "sfdx force:org:display --verbose";
+  public static async getCurrentOrg() {
+    const displayOrgCommand = "sfdx force:org:display";
     const displayResult = await execSfdxJson(displayOrgCommand, this, {
       fail: false,
-      output: true,
+      output: false,
     });
-    if (displayResult.id && type === "scratch" && displayResult.scratchOrg) {
-      return displayResult;
-    } else if (displayResult.id && type === "any") {
-      return displayResult;
+    if (displayResult?.result?.id) {
+      return displayResult.result;
     }
     return null;
   }
@@ -404,6 +402,12 @@ class MetadataUtils {
     const orgListResult = await execSfdxJson("sfdx force:org:list", this);
     if (type === "any") {
       return orgListResult?.result || [];
+    } else if (type === "sandbox") {
+      return (
+        orgListResult?.result?.nonScratchOrgs?.filter((org: any) => {
+          return org.loginUrl.includes("--") || org.loginUrl.includes("test.salesforce.com");
+        }) || []
+      );
     } else if (type === "scratch") {
       return (
         orgListResult?.result?.scratchOrgs?.filter((org: any) => {