+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [2.76.0] 2022-04-03
+
+- New command **sfdx**hardis:org:user:activateinvalid** to activate invalid emails in sandbox
+- Fix CI org authentication in case the default username is not the org that we want to configure
+- Bypass error with force:source:legacy:pull / push
+- hardis:work:save : Propose to manually commit files
+- Fix hardis:org:select alias & user config
+- Colorize command lines in logs
+- Enhance new task with sandbox (not fully stable yet)
+- New deployTips
+  - Please choose a different name
+
 ## [2.75.0] 2022-03-28
 
 - Property `availableTargetBranches` can be defined in `.sfdx-hardis.yml` to list the possible target branches for merge requests